Moved more to CUDA
Moved the CPU work done to select the best guess from the scores produced by the GPU over to the GPU.

Now the CUDA kernel scores everything, then does per-block reductions to find the best local guess. A second kernel comes behind and finishes the reduction to a single best guess that fits the original criteria.

This a more than 100% speedup over what was there before, depending on problem size and algo.