Memory improvement
Reduce the capacity of the possible solutions vector after removing impossible solutions. The number of remaining codewords is a tiny fraction of the original, yet the capacity was kept and even copied down to all levels in the strat. Ran out of memory at 15% done on MostParts 8p7c w/ 16GB RAM before this.