Refactoring, cleanup naming, simplify template parameters, place all strategy constants in a common place
The upshot of this is that most of the Strategy templating changed to pass down a StrategyConfig (or some variant of that) which holds pin count, color count, algo, subset size, etc.

Also unified the kernel args on the same concept.

It's easy to experiment w/ different launch parameters now.