Pick CUDA launch config based on subset size type
Subset sizes are counted up and used to form scores and it's the subset sizes that dictate how much shared mem is needed for the kernel.

Select the best subset size type to use based on the PS size at launch, with everything specialized so it all compiles away.

Further, pick the best launch config based on that subset size.

Substantial win across all games.

There's a histogram of which types get used in the game output now, which is pretty interesting. Focusing opts on 1-byte sizes is good, as is focusing on launch perf.