Coalesce PS on the GPU
Note: I gave up on keeping Metal working with this commit. I'll come back to it later.

Pass PS over to the GPU as Codeword<p,c>. Removes a lot of the work to get them over there and is a nice perf bump all around. Also sets us up to push the PS filtering to the GPU.