Cleanup AC setup
Just a little cleanup of setting up the All Codewords set. Makes the interface similar to how we get PS and UC to the GPU.

AC is still split arrays on the device. This performs better than having them interleaved, which is reasonable. Likely PS would perform better split as well, but it doesn't notice right now since a) it's usually quite small and b) it's a big win to get it over there as one block. Could split on-device, but likely not worth it for small sets.