New algorithm on GPU
This is a proof-of-concept which is a mess, all hacked up and janky, but it works for Knuth's algo and is faster than the old way. I'll start migrating to this and get it working for the other algos and strategy generation afterwards.

The old stuff was an evolution of a CPU-based serial implementation which put the big work on the GPU, but left a lot done on the CPU. While more could be moved to the GPU, the overall orginization meant that there was no opprotunity for batching.

This new impl plays all games at once with significant opportunity to batch up similar work in multiple places. This yields better GPU utilization and faster runs. It's already ~20% faster for Knuth 8p5c and I've done hardly any optimization work yet.

I used Thrust for the first time on this. Not sure yet if I like it or not, but it does make rapid prototyping easier.