Symmetry and Case Equivalence
Cleanup code to predict the size of the reduced AC. Added a doc with an explanation of the math that goes into this, the rationale, and some Python code to compute the result for all needed cases. The results are cached in a JSON file which is now read at runtime and the prediction is verified for every run.