CE tuning and new results
Only use the CE opt for GPU games > 400,000 codewords. 256 for CPU games. Sub-second games endup with a min time of the overhead to build the ACr sets, so dropping the opt for "smaller" games.

Minor cleanup here and there.

Doc updates for the CE opt on GPU.

Added new results for 6p15c and 7p11c.